requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.44.0
django-cors-headers==3.3.0
django-crispy-forms==1.10.0
django-extensions==2.2.5
django-jet==1.0.8
django-model-utils==4.0.0
django-phonenumber-field==5.0.0
django-storages==1.11.1
django-templated-mail==1.1.1
django-utils-six==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
drf-yasg==1.21.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jet==1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.1.0
packaging==21.3
phonenumbers==8.12.25
Pillow
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==1.5
python-decouple==3.4
python3-openid==3.2.0
pytz==2022.6
razorpay==1.2.0
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.2
whitenoise==5.2.0
